Update the RichTextEditor
