+++ packages/react-components/src/components/styles/RichTextEditor.styles.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * @private
+ */
+export const richTextEditorStyle = {
+  border: 'none',
+  overflow: 'auto',
+  padding: '10px',
+  outline: 'none',
+  bottom: '0',
+  minHeight: '2.25rem',
+  maxHeight: '8.25rem'
+};
